chore(developer): refactor to place keyboard info message unit tests in correct test file
